CyberSaathi_BackEnd/requirements.txt
# Core LangChain
langchain==0.3.13
langchain-community==0.3.13
langchain-google-genai==2.0.8
langgraph==0.2.60
langchain-huggingface==0.1.2

# Vector Store - ChromaDB (no database server needed!)
# Using latest version for better Windows compatibility
chromadb>=0.5.0

# Document Loaders
pypdf==5.1.0
docx2txt==0.8
python-docx==1.1.2

# API
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4

# Search
tavily-python==0.5.0

# Utilities
python-dotenv==1.0.1
sentence-transformers==3.3.1
tabulate==0.9.0
